added hyabusa2pds4gen tests
