+++ isis/src/hayabusa2/apps/hyb2pds4gen/hyb2pds4gen.h
@@ -1,0 +1,12 @@
+#ifndef hyb2pds4gen_h
+#define hyb2pds4gen_h 
+
+#include "UserInterface.h"
+#include "Cube.h"
+
+namespace Isis {
+  extern void hyb2pds4gen(UserInterface &ui);
+  extern void hyb2pds4gen(Cube *iCube, UserInterface &ui);
+}
+
+#endif
